% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_helpers.R
\name{calculate_absolute_baseline_counterfactual_distribution}
\alias{calculate_absolute_baseline_counterfactual_distribution}
\title{Calculate absolute baseline-counterfactual distribution
Calculates only the absolute numbers in the scenarios. The idea being that this will
be required for the calculation of the relative, but not vice-versa.}
\usage{
calculate_absolute_baseline_counterfactual_distribution(m, d_base, d_counter)
}
\arguments{
\item{m}{the model. Must contain age, sex, this_state, next_state, and other predictor variables}

\item{d_base}{The baseline dataset. Could be observed data from a specific wave, or purely hypothetical data}

\item{d_counter}{The counterfactual dataset. Could be a modification of d_base, and/or purely hypothetical data}
}
\value{
a tibble object with the specified values
}
\description{
Calculate absolute baseline-counterfactual distribution
Calculates only the absolute numbers in the scenarios. The idea being that this will
be required for the calculation of the relative, but not vice-versa.
}
