% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_helpers.R
\name{calculate_baseline_counterfactual_distribution}
\alias{calculate_baseline_counterfactual_distribution}
\title{Calculate baseline and counterfactual state distributions
Given a model, a baseline dataset, and a counterfactual dataset, calculate the estimated distribution of
the population in each economic (in)activity state in the next wave}
\usage{
calculate_baseline_counterfactual_distribution(
  m,
  d_base,
  d_counter,
  returnType = c("absolute", "relative")
)
}
\arguments{
\item{m}{the model. Must contain age, sex, this_state, next_state, and other predictor variables}

\item{d_base}{The baseline dataset. Could be observed data from a specific wave, or purely hypothetical data}

\item{d_counter}{The counterfactual dataset. Could be a modification of d_base, and/or purely hypothetical data. Could also be a named list of data frames}

\item{returnType}{Either absolute, relative, or both. Whether to return the absolute numbers in each state under each scenario, or
the change from the baseline scenario}
}
\value{
a tibble object with the specified values
}
\description{
Calculate baseline and counterfactual state distributions
Given a model, a baseline dataset, and a counterfactual dataset, calculate the estimated distribution of
the population in each economic (in)activity state in the next wave
}
