% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukhls_data_extractors.R
\name{convert_varname_selection_to_regex}
\alias{convert_varname_selection_to_regex}
\title{Convert variable names to regex search string}
\usage{
convert_varname_selection_to_regex(varnames, level = "individual")
}
\arguments{
\item{varnames}{list of variables except for pidp, and without {w}_ prefix part}
}
\value{
out : a string for passing to the matches function in the col_select parameter
in haven::read_dta
}
\description{
Convert variable names to regex search string
}
