% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukhls_data_extractors.R
\name{extract_eq_income_and_num_dependents}
\alias{extract_eq_income_and_num_dependents}
\title{Extract equivalised monthly income and number of dependent children from hhresp file}
\usage{
extract_eq_income_and_num_dependents(file_location)
}
\arguments{
\item{file_location}{The location of a household level datafile of type hhresp}
}
\value{
dta_hh. A dataframe with variables in long format and with columns named in a
more informative way
}
\description{
Extract equivalised monthly income and number of dependent children from hhresp file
}
