% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukhls_data_reshapers.R
\name{extract_ind_from_waves}
\alias{extract_ind_from_waves}
\title{Extract individual level data from all waves}
\usage{
extract_ind_from_waves(
  varnames,
  vartypes,
  strict = FALSE,
  verbose = FALSE,
  waves = "all",
  base_dir_location = "big_data/UKDA-6614-stata/stata/stata13_se/ukhls"
)
}
\arguments{
\item{varnames}{The names of the variables to extract}

\item{vartypes}{The types of the variables to extract (labels or values)}

\item{strict}{If TRUE, the function will fail if not all variables are found in all waves}

\item{verbose}{Whether to report information about progress or not}

\item{waves}{Either the keyword 'all', or a vector of letters corresponding to which waves to include}

\item{base_dir_location}{Where the collection of _indresp files can be found. Do not change unless necessary}
}
\value{
Dataframe with wave, pid and other variables as columns (i.e. pre-widened)
}
\description{
Extract individual level data from all waves
}
