% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukhls_data_extractors.R
\name{extract_vars_and_make_long}
\alias{extract_vars_and_make_long}
\title{Extract individual level variables and convert to long format}
\usage{
extract_vars_and_make_long(
  dta,
  varnames,
  extract_what = "labels",
  level = "individual",
  verbose = TRUE
)
}
\arguments{
\item{dta}{a dataframe from ukhls}

\item{varnames}{character vector, length K: the names of the variables to extract (apart from
pidp which is implicit)}

\item{extract_what}{a vector of either length 1 or length K, each of whose elements must be one of
'labels', 'values', or 'both'}

\item{verbose}{boolean. If true then time taken to extract is reported}
}
\value{
dataframe with only selected variables, and in long format
}
\description{
Extract individual level variables and convert to long format
}
