% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukhls_helpers.R
\name{get_col_labels}
\alias{get_col_labels}
\title{Get full column labels
If the data are}
\usage{
get_col_labels(dta)
}
\arguments{
\item{dta}{A dataframe containing some columns which are haven_labelled}
}
\value{
out
A vector with the column labels
}
\description{
Get full column labels
If the data are
}
