% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukhls_helpers.R
\name{get_labels_for_col}
\alias{get_labels_for_col}
\title{get labels for column}
\usage{
get_labels_for_col(x)
}
\arguments{
\item{x}{a haven_labelled column passed as a simple vector}
}
\value{
an object showing the value and labels lookup for the haven_labelled column
}
\description{
get labels for column
}
