% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_helpers.R
\name{make_tabular_summary}
\alias{make_tabular_summary}
\title{Produce a tabular summary of results from plot_scenario_comparisons}
\usage{
make_tabular_summary(x, title = NULL, subtitle = NULL, outputType = "standard")
}
\arguments{
\item{x}{A valid output from plot_scenario_comparisons}

\item{title}{the title to give the table}

\item{subtitle}{the subtitle to give the table}
}
\value{
the summary output
}
\description{
Produce a tabular summary of results from plot_scenario_comparisons
}
