% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_helpers.R
\name{plot_scenario_comparisons}
\alias{plot_scenario_comparisons}
\title{Plot scenario comparison information
Given the output of calculate_baseline_counterfactual_distribution, produce nice plots to visually summarise the information}
\usage{
plot_scenario_comparisons(
  dists,
  title = "Default Title",
  plotWhat = "absolute",
  simplify = FALSE,
  polar = FALSE,
  dynamic = FALSE
)
}
\arguments{
\item{dists}{an object created by calculate_baseline_counterfactual_distribution, which should be of class distFrame}

\item{title}{What to put on the title of the figure}

\item{plotWhat}{Whether to plot absolute, relative, or both}

\item{simplify}{Whether to reduce from seven mutually exclusive states to 2 (Will only work with absolute)}

\item{polar}{Whether to show distribution as polar cooardinates}

\item{dynamic}{Whether to show a static ggplot2 chart (default) or a plotly interactive chart}
}
\description{
Plot scenario comparison information
Given the output of calculate_baseline_counterfactual_distribution, produce nice plots to visually summarise the information
}
