% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukhls_data_extractors.R
\name{read_and_slim_data}
\alias{read_and_slim_data}
\title{Read individual and household level file and extract variables}
\usage{
read_and_slim_data(
  file_location,
  varnames,
  extract_what,
  verbose = FALSE,
  level = "individual"
)
}
\arguments{
\item{file_location}{string. the name of the file}

\item{varnames}{string vector of length K. The names of the variables (except pidp which is implicit)
excluding the {w}_ prefix expected for most variables in ukhls}

\item{extract_what}{string vector of length 1 or K. Each element can be one of
'values', 'labels', or 'both}

\item{verbose}{boolean. Whether to report which files are being ingested and how long
the process is taking.}
}
\description{
Read individual and household level file and extract variables
}
