% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukhls_helpers.R
\name{return_labels_as_factors}
\alias{return_labels_as_factors}
\title{Return labels as factors}
\usage{
return_labels_as_factors(dta)
}
\arguments{
\item{dta}{The dataset containing some columns which are haven labelled}
}
\description{
If the dataframe contains haven labelled variables, this turns all those
columns that are haven labelled into more straightforward characters
}
