% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukhls_data_reshapers.R
\name{simplify_econ_status_categories}
\alias{simplify_econ_status_categories}
\title{Simplify economic status categories
Convert economic status categories to simplified and standardised groupings}
\usage{
simplify_econ_status_categories(
  original,
  level = 3,
  newName = "econ_status",
  missingAsNa = TRUE
)
}
\arguments{
\item{original}{Usually the variable jbstat. Can be either a data.frame of 1 column, or a character vector}

\item{level}{The level of aggregation to perform. Options are
\enumerate{
\item 2 levels (Active, Inactive)
\item 3 levels (Employed, Unemployed, Inactive)
\item 7 levels (2 active levels; 5 inactive levels)
\item Multiple levels - effectively just tidying original variable labels, rather than reaggregating
}}

\item{newName}{The name of the new variable to create}

\item{missingAsNa}{Whether to convert the category 'Missing' to NA (default TRUE)}
}
\value{
revised The revised categorisations. Just a character vector.
}
\description{
Simplify economic status categories
Convert economic status categories to simplified and standardised groupings
}
