% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukhls_data_reshapers.R
\name{smartly_widen_ind_dataframe}
\alias{smartly_widen_ind_dataframe}
\title{Intelligently widen long individual level dataframes}
\usage{
smartly_widen_ind_dataframe(df, varnames, vartypes, verbose = FALSE)
}
\arguments{
\item{df}{the dataframe in long format. Must contain wave and pidp variables,
and columns called variable and value}

\item{varnames}{the variable names in the variable column to work on}

\item{vartypes}{The type of variable to convert the variables to (labels or values)}

\item{verbose}{Boolean. Whether to report on progress while executing}
}
\value{
out a widened dataframe with pipd and wave as columns, followed by additional columns for each
variable specified in varnames and found in df
}
\description{
Intelligently widen long individual level dataframes
}
