% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukhls_helpers.R
\name{standardise_scores}
\alias{standardise_scores}
\title{Standardise scores}
\usage{
standardise_scores(x)
}
\arguments{
\item{x}{A vector of type numeric/double}
}
\value{
out: A vector of the same length as x, but with scores standardised
}
\description{
Standardise scores
}
